Don't look for activity in application
In this [StackOverflow question](https://stackoverflow.com/questions/62213443/finding-cause-in-leakcanary-leak-trace) I noticed this bit:

```
D LeakCanary: ├─ <>.application.HomeApplication
D LeakCanary: │    Leaking: NO (Application is a singleton)
D LeakCanary: │    HomeApplication does not wrap an activity context
```

The comment `HomeApplication does not wrap an activity context` doesn't make sense for an application, it's meant for ContextWrapper instances that may wrap an activity context.

This fixes that, and optimizes things a bit so that we don't keep looking up the full class hierarchy.